Renames package org.valiktor.springframework.handler to org.valiktor.springframework.http
