+++ valiktor-samples/valiktor-sample-spring-boot-2-webmvc/src/test/kotlin/org/valiktor/sample/EmployeeControllerTestFixture.kt
@@ -16,10 +16,10 @@
 
 package org.valiktor.sample
 
-import org.valiktor.springframework.handler.UnprocessableEntity
-import org.valiktor.springframework.handler.ValidationConstraint
-import org.valiktor.springframework.handler.ValidationError
-import org.valiktor.springframework.handler.ValidationParam
+import org.valiktor.springframework.http.UnprocessableEntity
+import org.valiktor.springframework.http.ValidationConstraint
+import org.valiktor.springframework.http.ValidationError
+import org.valiktor.springframework.http.ValidationParam
 import java.time.LocalDate
 import java.time.Month
 import java.util.Locale
