+++ valiktor-spring/valiktor-spring-boot-autoconfigure/src/main/kotlin/org/valiktor/springframework/boot/autoconfigure/ValiktorWebMvcAutoConfiguration.kt
@@ -21,10 +21,10 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnClass
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
-import org.valiktor.springframework.handler.ValiktorExceptionHandler
-import org.valiktor.springframework.handler.webmvc.ConstraintViolationExceptionHandler
-import org.valiktor.springframework.handler.webmvc.InvalidFormatExceptionHandler
-import org.valiktor.springframework.handler.webmvc.MissingKotlinParameterExceptionHandler
+import org.valiktor.springframework.http.ValiktorExceptionHandler
+import org.valiktor.springframework.http.webmvc.ConstraintViolationExceptionHandler
+import org.valiktor.springframework.http.webmvc.InvalidFormatExceptionHandler
+import org.valiktor.springframework.http.webmvc.MissingKotlinParameterExceptionHandler
 
 /**
  * Represents the SpringBoot Auto Configuration for Spring WebFlux exception handlers
@@ -34,7 +34,7 @@ import org.valiktor.springframework.handler.webmvc.MissingKotlinParameterExcepti
  */
 @Configuration
 @ConditionalOnClass(name = ["org.springframework.web.servlet.DispatcherServlet"])
-@ConditionalOnBean(type = ["org.valiktor.springframework.handler.ValiktorExceptionHandler"])
+@ConditionalOnBean(type = ["org.valiktor.springframework.http.ValiktorExceptionHandler"])
 class ValiktorWebMvcAutoConfiguration {
 
     /**
