+++ valiktor-spring/valiktor-spring-boot-autoconfigure/src/test/kotlin/org/valiktor/springframework/boot/autoconfigure/ValiktorWebMvcAutoConfigurationTest.kt
@@ -25,10 +25,10 @@ import org.springframework.boot.test.context.runner.ApplicationContextRunner
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.web.servlet.DispatcherServlet
-import org.valiktor.springframework.handler.ValiktorExceptionHandler
-import org.valiktor.springframework.handler.webmvc.ConstraintViolationExceptionHandler
-import org.valiktor.springframework.handler.webmvc.InvalidFormatExceptionHandler
-import org.valiktor.springframework.handler.webmvc.MissingKotlinParameterExceptionHandler
+import org.valiktor.springframework.http.ValiktorExceptionHandler
+import org.valiktor.springframework.http.webmvc.ConstraintViolationExceptionHandler
+import org.valiktor.springframework.http.webmvc.InvalidFormatExceptionHandler
+import org.valiktor.springframework.http.webmvc.MissingKotlinParameterExceptionHandler
 import kotlin.test.Test
 
 class ValiktorWebMvcAutoConfigurationTest {
