+++ valiktor-spring/valiktor-spring/src/main/kotlin/org/valiktor/springframework/http/webmvc/MissingKotlinParameterExceptionHandler.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.valiktor.springframework.handler.webmvc
+package org.valiktor.springframework.http.webmvc
 
 import com.fasterxml.jackson.module.kotlin.MissingKotlinParameterException
 import org.springframework.http.ResponseEntity
