+++ valiktor-spring/valiktor-spring/src/test/kotlin/org/valiktor/springframework/http/webflux/ReactiveConstraintViolationExceptionHandlerJsonTest.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.valiktor.springframework.handler.webflux
+package org.valiktor.springframework.http.webflux
 
 import org.springframework.http.HttpHeaders.ACCEPT_LANGUAGE
 import org.springframework.http.HttpHeaders.LOCATION
