+++ valiktor-spring/valiktor-spring/src/test/kotlin/org/valiktor/springframework/http/webmvc/ConstraintViolationExceptionHandlerJsonTest.kt
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-package org.valiktor.springframework.handler.webmvc
+package org.valiktor.springframework.http.webmvc
 
 import org.springframework.http.HttpHeaders.ACCEPT_LANGUAGE
 import org.springframework.http.HttpHeaders.LOCATION
-import org.springframework.http.MediaType.APPLICATION_XML
+import org.springframework.http.MediaType.APPLICATION_JSON
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post
 import org.springframework.test.web.servlet.result.MockMvcResultHandlers.log
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers.content
@@ -28,10 +28,10 @@ import java.util.Locale
 import kotlin.test.BeforeTest
 import kotlin.test.Test
 
-class MissingKotlinParameterExceptionHandlerXmlTest {
+class ConstraintViolationExceptionHandlerJsonTest {
 
     private val mockMvc = ExceptionHandlerFixture.mockMvc
-    private val xml = ExceptionHandlerFixture.XML
+    private val json = ExceptionHandlerFixture.JSON
 
     @BeforeTest
     fun setUp() {
@@ -42,9 +42,9 @@ class MissingKotlinParameterExceptionHandlerXmlTest {
     fun `should return 201`() {
         mockMvc
             .perform(post("/employees")
-                .accept(APPLICATION_XML)
-                .contentType(APPLICATION_XML)
-                .content(xml.payloadEmployeeValid()))
+                .accept(APPLICATION_JSON)
+                .contentType(APPLICATION_JSON)
+                .content(json.payloadEmployeeValid()))
             .andExpect(status().isCreated)
             .andExpect(header().string(LOCATION, "http://localhost/employees/1"))
             .andExpect(content().bytes(ByteArray(0)))
@@ -55,12 +55,12 @@ class MissingKotlinParameterExceptionHandlerXmlTest {
     fun `should return 422 with default locale`() {
         mockMvc
             .perform(post("/employees")
-                .accept(APPLICATION_XML)
-                .contentType(APPLICATION_XML)
-                .content(xml.payloadEmployeeNullName()))
+                .accept(APPLICATION_JSON)
+                .contentType(APPLICATION_JSON)
+                .content(json.payloadEmployeeInvalid()))
             .andExpect(status().isUnprocessableEntity)
-            .andExpect(content().contentTypeCompatibleWith(APPLICATION_XML))
-            .andExpect(content().xml(xml.payload422NullName(Locale.ENGLISH)))
+            .andExpect(content().contentTypeCompatibleWith(APPLICATION_JSON))
+            .andExpect(content().json(json.payload422(Locale.ENGLISH)))
             .andDo(log())
     }
 
@@ -68,13 +68,13 @@ class MissingKotlinParameterExceptionHandlerXmlTest {
     fun `should return 422 with locale en`() {
         mockMvc
             .perform(post("/employees")
-                .accept(APPLICATION_XML)
+                .accept(APPLICATION_JSON)
                 .header(ACCEPT_LANGUAGE, "en")
-                .contentType(APPLICATION_XML)
-                .content(xml.payloadEmployeeNullName()))
+                .contentType(APPLICATION_JSON)
+                .content(json.payloadEmployeeInvalid()))
             .andExpect(status().isUnprocessableEntity)
-            .andExpect(content().contentTypeCompatibleWith(APPLICATION_XML))
-            .andExpect(content().xml(xml.payload422NullName(Locale.ENGLISH)))
+            .andExpect(content().contentTypeCompatibleWith(APPLICATION_JSON))
+            .andExpect(content().json(json.payload422(Locale.ENGLISH)))
             .andDo(log())
     }
 
@@ -82,13 +82,13 @@ class MissingKotlinParameterExceptionHandlerXmlTest {
     fun `should return 422 with locale pt_BR`() {
         mockMvc
             .perform(post("/employees")
-                .accept(APPLICATION_XML)
+                .accept(APPLICATION_JSON)
                 .header(ACCEPT_LANGUAGE, "pt-BR")
-                .contentType(APPLICATION_XML)
-                .content(xml.payloadEmployeeNullName()))
+                .contentType(APPLICATION_JSON)
+                .content(json.payloadEmployeeInvalid()))
             .andExpect(status().isUnprocessableEntity)
-            .andExpect(content().contentTypeCompatibleWith(APPLICATION_XML))
-            .andExpect(content().xml(xml.payload422NullName(Locale("pt", "BR"))))
+            .andExpect(content().contentTypeCompatibleWith(APPLICATION_JSON))
+            .andExpect(content().json(json.payload422(Locale("pt", "BR"))))
             .andDo(log())
     }
 }
