+++ valiktor-spring/valiktor-spring/src/test/kotlin/org/valiktor/springframework/http/webmvc/ExceptionHandlerFixture.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.valiktor.springframework.handler.webmvc
+package org.valiktor.springframework.http.webmvc
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.databind.SerializationFeature
@@ -39,9 +39,9 @@ import org.valiktor.functions.isBetween
 import org.valiktor.functions.isEmail
 import org.valiktor.functions.isEqualTo
 import org.valiktor.springframework.config.ValiktorConfiguration
-import org.valiktor.springframework.handler.DefaultValiktorExceptionHandler
-import org.valiktor.springframework.handler.UnprocessableEntity
-import org.valiktor.springframework.handler.ValiktorExceptionHandler
+import org.valiktor.springframework.http.DefaultValiktorExceptionHandler
+import org.valiktor.springframework.http.UnprocessableEntity
+import org.valiktor.springframework.http.ValiktorExceptionHandler
 import org.valiktor.validate
 import java.math.BigDecimal
 import java.text.SimpleDateFormat
